refactor(migration): move CodernityDB->SQLite migration to a standalone one-time script (REFACTOR-01)
The CodernityDB->SQLite migration was one-time upgrade code living inline in
couchpotato/runner.py and inside couchpotato/core/migration/, pulling
CodernityDB into every server process's import graph even on
already-migrated installs. Move the migration logic (codernity_to_sqlite,
fix_indexes, rebuild_buckets) into a standalone, directly-runnable
scripts/migrate_codernity_to_sqlite.py; runner.py now only detects a legacy
database and runs the script ONCE as a subprocess, aborting startup (instead
of silently creating a fresh, empty database) if it fails. clean_orphans.py
and fix_release_quality.py stay put since they run on every startup, not
just during migration. CodernityDB itself is untouched.
