+++ tests/unit/test_database_class.py
@@ -481,69 +481,3 @@ class TestCleanOrphansEdgeCases:
 
         removed = clean_orphaned_movies(db)
         assert removed == 0  # Should keep because it has original_title
-
-
-class TestRebuildBucketsEdgeCases:
-    """Additional edge cases for rebuild_after_migration."""
-
-    def test_handles_corrupt_entry(self, tmp_path):
-        """Rebuild should skip corrupt entries gracefully."""
-        from CodernityDB.database import Database
-        from couchpotato.core.migration.rebuild_buckets import rebuild_after_migration
-
-        db_path = str(tmp_path / 'db')
-        db = Database(db_path)
-        db.create()
-
-        # Insert valid records
-        for i in range(5):
-            db.insert({'index': i})
-
-        db.close()
-
-        # Reopen and rebuild
-        db2 = Database(db_path)
-        rebuild_after_migration(db2, db_path)
-
-        # Should still work
-        count = sum(1 for _ in db2.all('id'))
-        assert count == 5
-        db2.close()
-
-
-class TestFixIndexesEdgeCases:
-    """Additional edge cases for fix_index_files."""
-
-    @pytest.fixture
-    def db_dir(self, tmp_path):
-        indexes_dir = tmp_path / '_indexes'
-        indexes_dir.mkdir()
-        return tmp_path, indexes_dir
-
-    def test_handles_permission_error(self, db_dir, monkeypatch):
-        """Should handle unreadable files gracefully."""
-        from couchpotato.core.migration.fix_indexes import fix_index_files
-
-        db_path, indexes_dir = db_dir
-        # This test would need actual permission changes which may not work
-        # in all environments, so we just verify the function doesn't crash
-        result = fix_index_files(str(db_path))
-        assert result >= 0
-
-    def test_handles_complex_md5_expression(self, db_dir):
-        """Should fix complex expressions inside md5()."""
-        from couchpotato.core.migration.fix_indexes import fix_index_files
-
-        db_path, indexes_dir = db_dir
-        content = '''from hashlib import md5
-def make_key(data):
-    return md5(data.get("field", "") + "_suffix").hexdigest()
-'''
-        path = indexes_dir / 'complex.py'
-        path.write_text(content)
-
-        result = fix_index_files(str(db_path))
-        assert result == 1
-
-        fixed = path.read_text()
-        assert '_to_bytes' in fixed
